Resolve conflicts between normal shortcuts and QuickAccessors.
Now both shortcuts pages see all ActionCollections but filter out the irrelevant ones.
(Should we remove the arguments from the constructor?) And resolve conflicts using the
existing mechanisms. A signal is now emitted when an ActionCollection's action is changed,
and the ShortcutsModel listens for this to update shortcuts changed from another page.
We can't customise the KDE page, so shortcuts overwritten there are currently not even
warned about.